requirements.txt
click==8.1.3
colorama==0.4.6
Flask==2.2.2
gitdb==4.0.9
GitPython==3.1.29
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
smmap==5.0.0
Werkzeug==2.2.2
zipp==3.10.0

bottle~=0.12.23
jira~=3.4.1